Day6 puzzle 2 solution with kotlin.

+++ day6/src/main/java/fi/viware/aoc2023/day6/Boat.kt
@@ -2,7 +2,7 @@ package fi.viware.aoc2023.day6
 
 class Boat {
 
-    fun sailDistance(pressTime:Int, raceDuration: Int): Int{
+    fun sailDistance(pressTime:Long, raceDuration: Long): Long{
 
         val speed = pressTime
         val retDist = (raceDuration - pressTime)*speed
