+++ day6/src/main/java/fi/viware/aoc2023/day6/Race.kt
@@ -1,6 +1,6 @@
 package fi.viware.aoc2023.day6
 
-class Race(val time: Int, val distance: Int) {
+class Race(val time: Long, val distance: Long) {
 
     override fun toString(): String {
         return "time: $time, distance $distance"
